Give intervention details their own presenter
This was initially used for the Search Results page, but we can re-use
it for the Intervention Details page.
