+++ server/routes/findInterventions/interventionDetailsPresenter.ts
@@ -1,0 +1,75 @@
+import { Eligibility, Intervention } from '../../services/interventionsService'
+import { SummaryListItem } from '../../utils/summaryList'
+import utils from '../../utils/utils'
+
+export default class InterventionDetailsPresenter {
+  constructor(private readonly intervention: Intervention) {}
+
+  get title(): string {
+    return this.intervention.title
+  }
+
+  get href(): string {
+    return `/find-interventions/intervention/${this.intervention.id}`
+  }
+
+  get body(): string {
+    return this.intervention.description
+  }
+
+  get summary(): SummaryListItem[] {
+    return [
+      {
+        key: 'Type',
+        lines: ['Dynamic Framework'],
+        isList: false,
+      },
+      {
+        key: 'Location',
+        lines: [this.intervention.pccRegions.map(region => region.name).join(', ')],
+        isList: false,
+      },
+      {
+        key: 'Criminogenic needs',
+        lines: [utils.convertToProperCase(this.intervention.serviceCategory.name)],
+        isList: false,
+      },
+      {
+        key: 'Provider',
+        lines: [this.intervention.serviceProvider.name],
+        isList: false,
+      },
+      {
+        key: 'Age group',
+        lines: [InterventionDetailsPresenter.ageGroupDescription(this.intervention.eligibility)],
+        isList: false,
+      },
+      {
+        key: 'Sex',
+        lines: [InterventionDetailsPresenter.sexDescription(this.intervention.eligibility)],
+        isList: false,
+      },
+    ]
+  }
+
+  static ageGroupDescription(eligibility: Eligibility): string {
+    if (eligibility.maximumAge === null) {
+      return `${eligibility.minimumAge}+`
+    }
+    return `${eligibility.minimumAge}–${eligibility.maximumAge}`
+  }
+
+  static sexDescription(eligibility: Eligibility): string {
+    if (eligibility.allowsMale && eligibility.allowsFemale) {
+      return 'Male and female'
+    }
+    if (eligibility.allowsMale) {
+      return 'Male'
+    }
+    if (eligibility.allowsFemale) {
+      return 'Female'
+    }
+
+    return ''
+  }
+}
